Rename: move config into services folder
