+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import Greeting from './Greeting';
 import Board from './Board';
-import Config from './config';
+import Config from './services/config';
 import Api from './api';
 
 const setTitle = (title) => document.title = title;
